requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
django-allauth
